Unit testing

+++ cylance_protect/icon_cylance_protect/actions/delete_asset/action.py
@@ -23,38 +23,44 @@ class DeleteAsset(insightconnect_plugin_runtime.Action):
         valid_devices = []
         invalid_devices = []
         for agent in agents:
-            if validators.ipv4(agent):
-                agent = find_agent_by_ip(self.connection, agent)
-
-            device_obj = self.connection.client.get_agent_details(agent)
-
-            if device_obj:
-                if whitelist:
-                    matches = find_in_whitelist(device_obj, whitelist)
-                    if matches:
-                        invalid_devices.append(agent)
+            try:
+                if validators.ipv4(agent):
+                    agent = find_agent_by_ip(self.connection, agent)
+                device_obj = self.connection.client.get_agent_details(agent)
+                if device_obj:
+                    if whitelist:
+                        matches = find_in_whitelist(device_obj, whitelist)
+                        if matches:
+                            invalid_devices.append(agent)
+                        else:
+                            valid_devices.append(agent)
+                            valid_ids = self.add_to_valid_devices(device_obj, valid_ids)
                     else:
                         valid_devices.append(agent)
                         valid_ids = self.add_to_valid_devices(device_obj, valid_ids)
                 else:
-                    valid_devices.append(agent)
-                    valid_ids = self.add_to_valid_devices(device_obj, valid_ids)
-            else:
-                self.connection.logger.info("NOT FOUND %s" % agent)
+                    invalid_devices.append(agent)
+            except PluginException:
                 invalid_devices.append(agent)
+
+        if not valid_ids:
+            raise PluginException(
+                cause="No valid devices to delete.",
+                assistance=f"Be sure that the devices exist in Cylance and are not part of the whitelist."
+            )
 
         payload = {"device_ids": valid_ids}
         success = self.connection.client.delete_devices(payload)
         if not success:
             raise PluginException(
                 cause="One of the devices failed to delete.",
-                assistance=f"Example assistance"
+                assistance=f"A valid agent deletion may have failed, check your Cylance console."
             )
-            return {Output.SUCCESS: False}
 
         return {Output.SUCCESS: True, Output.DELETED: valid_devices, Output.NOT_DELETED: invalid_devices}
 
-    def add_to_valid_devices(self, device_obj, valid_ids):
+    @staticmethod
+    def add_to_valid_devices(device_obj, valid_ids):
         device_id = device_obj.get('id')
         device_id.replace('-', '').upper()
         valid_ids.append(device_id)
